got the bare bones functionalities done
